png2ico/requirements.txt
Pillow==10.4.0
rich==13.8.1
